Code cleanup, references made, consts added, on some parts of the code. DLL debug options, deltatime is now double (more precise)

+++ Rendering/src/Rendering/Context/OpenGL/GLFWDevice.cpp
@@ -66,7 +66,7 @@ void GLFWDevice::CalculateDeltaTime() noexcept
     lastTime = currentTime;
 }
 
-float GLFWDevice::GetDeltaTime() const noexcept
+double GLFWDevice::GetDeltaTime() const noexcept
 {
     return deltaTime;
 }
